Refactor Makefile and add dashboard setup.py

+++ Makefile
@@ -13,11 +13,12 @@ CA_KEY_FILE_PATH := ca-key.pem
 CA_CERT_FILE_PATH := ca-cert.pem
 CA_SIGNING_KEY_FILE_PATH := ca-signing-key.pem
 
-.PHONY: all clean-lib test-lib package test-release release coverage lint autopep8
-.PHONY: container run-container release-container https-certificates ca-certificates
-.PHONY: profile dashboard clean-dashboard
+.PHONY: all https-certificates ca-certificates autopep8
+.PHONY: lib-clean lib-test lib-package lib-release-test lib-release lib-coverage lib-lint lib-profile
+.PHONY: container container-run container-release
+.PHONY: dashboard dashboard-clean dashboard-package
 
-all: clean-lib test-lib
+all: lib-clean lib-test
 
 autopep8:
 	autopep8 --recursive --in-place --aggressive proxy/*.py
@@ -28,6 +29,12 @@ autopep8:
 	autopep8 --recursive --in-place --aggressive dashboard/*.py
 	autopep8 --recursive --in-place --aggressive setup.py
 
+https-certificates:
+	# Generate server key
+	openssl genrsa -out $(HTTPS_KEY_FILE_PATH) 2048
+	# Generate server certificate
+	openssl req -new -x509 -days 3650 -key $(HTTPS_KEY_FILE_PATH) -out $(HTTPS_CERT_FILE_PATH)
+
 ca-certificates:
 	# Generate CA key
 	openssl genrsa -out $(CA_KEY_FILE_PATH) 2048
@@ -37,7 +44,7 @@ ca-certificates:
 	# Generated certificates are then signed with CA certificate / key generated above
 	openssl genrsa -out $(CA_SIGNING_KEY_FILE_PATH) 2048
 
-clean-lib:
+lib-clean:
 	find . -name '*.pyc' -exec rm -f {} +
 	find . -name '*.pyo' -exec rm -f {} +
 	find . -name '*~' -exec rm -f {} +
@@ -49,47 +56,47 @@ clean-lib:
 	rm -rf .pytest_cache
 	rm -rf .hypothesis
 
-clean-dashboard:
-	rm -rf public/dashboard
+lib-lint:
+	flake8 --ignore=W504 --max-line-length=127 proxy/ tests/ benchmark/ plugin_examples/ dashboard/dashboard.py setup.py
+	mypy --strict --ignore-missing-imports proxy/ tests/ benchmark/ plugin_examples/ dashboard/dashboard.py setup.py
 
-container:
-	docker build -t $(LATEST_TAG) -t $(IMAGE_TAG) .
+lib-test: lib-lint
+	python -m unittest discover
 
-coverage:
+lib-release-test: package
+	twine upload --verbose --repository-url https://test.pypi.org/legacy/ dist/*
+
+lib-coverage:
 	pytest --cov=proxy --cov-report=html tests/
 	open htmlcov/index.html
+
+lib-package: lib-clean
+	python setup.py sdist bdist_wheel
+
+lib-release: lib-package
+	twine upload dist/*
+
+lib-profile:
+	sudo py-spy -F -f profile.svg -d 3600 proxy.py
 
 dashboard:
 	pushd dashboard && npm run build && popd
 
-https-certificates:
-	# Generate server key
-	openssl genrsa -out $(HTTPS_KEY_FILE_PATH) 2048
-	# Generate server certificate
-	openssl req -new -x509 -days 3650 -key $(HTTPS_KEY_FILE_PATH) -out $(HTTPS_CERT_FILE_PATH)
+dashboard-clean:
+	if [[ -d public/dashboard ]]; then rm -rf public/dashboard; fi
 
-lint:
-	flake8 --ignore=W504 --max-line-length=127 proxy/ tests/ benchmark/ plugin_examples/ dashboard/dashboard.py setup.py
-	mypy --strict --ignore-missing-imports proxy/ tests/ benchmark/ plugin_examples/ dashboard/dashboard.py setup.py
+dashboard-package-clean:
+	pushd dashboard && rm -rf build && rm -rf dist && popd
 
-package: clean
-	python setup.py sdist bdist_wheel
+dashboard-package: dashboard-package-clean
+	pushd dashboard && python setup.py sdist bdist_wheel && popd
 
-profile:
-	sudo py-spy -F -f profile.svg -d 3600 proxy.py
+container:
+	docker build -t $(LATEST_TAG) -t $(IMAGE_TAG) .
 
-release: package
-	twine upload dist/*
-
-release-container:
+container-release:
 	docker push $(IMAGE_TAG)
 	docker push $(LATEST_TAG)
 
-run-container:
+container-run:
 	docker run -it -p 8899:8899 --rm $(LATEST_TAG)
-
-test-lib: lint
-	python -m unittest discover
-
-test-release: package
-	twine upload --verbose --repository-url https://test.pypi.org/legacy/ dist/*
